Server/requirements.txt
Flask
Flask-Cors
Flask-Mail
Flask-Bcrypt
supabase
python-dotenv
email-validator
password-validator